requirements.txt
flask
bs4
mysql-connector-python
gunicorn
requests
lxml
matplotlib
